+++ src/bittytax/conv/parsers/ledgerlive.py
@@ -23,6 +23,9 @@ def parse_ledger_live(
     row_dict = data_row.row_dict
     data_row.timestamp = DataParser.parse_timestamp(row_dict["Operation Date"])
 
+    if "Status" in row_dict and row_dict["Status"] != "Confirmed":
+        return
+
     if row_dict["Operation Fees"]:
         fee_quantity = Decimal(row_dict["Operation Fees"])
         fee_asset = row_dict["Currency Ticker"]
@@ -90,6 +93,27 @@ def parse_ledger_live(
             row_dict["Operation Type"],
         )
 
+
+DataParser(
+    ParserType.WALLET,
+    "Ledger Live",
+    [
+        "Operation Date",
+        "Status",
+        "Currency Ticker",
+        "Operation Type",
+        "Operation Amount",
+        "Operation Fees",
+        "Operation Hash",
+        "Account Name",
+        "Account xpub",
+        "Countervalue Ticker",
+        "Countervalue at Operation Date",
+        "Countervalue at CSV Export",
+    ],
+    worksheet_name="Ledger",
+    row_handler=parse_ledger_live,
+)
 
 DataParser(
     ParserType.WALLET,